requirements.txt
fastapi
uvicorn[standard]
numpy==1.26.4
tensorflow
Pillow
python-multipart
ultralytics
sqlalchemy
psycopg2-binary
python-dotenv